Revert "Merge branch 'master' of https://github.com/krogvig/LightWay"
This reverts commit eaca65200faef4e564ee4615cc9917c5092934ef, reversing
changes made to 81c1b5336b971795e7f66b80baa036780998dd31.
