+++ app/build.gradle
@@ -28,13 +28,11 @@ dependencies {
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
+    //implementation 'com.google.android.gms:play-services-maps:12.2.21'
     implementation 'com.google.android.gms:play-services-maps:15.0.0'
     implementation 'com.google.android.gms:play-services-location:15.0.0'
     implementation 'com.google.android.gms:play-services-places:15.0.0'
     implementation 'com.google.android.gms:play-services-nearby:15.0.0'
-    implementation 'com.google.android.gms:play-services-nearby:15.0.0'
-    implementation 'com.google.android.gms:play-services-auth:15.0.0'
-    implementation 'com.android.volley:volley:1.0.0'
 
 
 }
