--- app/credentials.json
@@ -1,1 +1,0 @@
-{"web":{"client_id":"157575935545-1g43gu789i9tk1hv23k2qcn9n32p0ifj.apps.googleusercontent.com","project_id":"lightway-202411","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://accounts.google.com/o/oauth2/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"VUNXtuQhwmgYzE8tBYywM9_-"}}
