--- app/src/main/java/com/example/lightway/RegisterActivity.java
@@ -1,13 +1,0 @@
-package com.example.lightway;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class RegisterActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
-    }
-}
