+++ app/src/main/res/layout/activity_main.xml
@@ -6,43 +6,46 @@
     android:layout_height="match_parent"
     tools:context="com.example.lightway.MainActivity">
 
-    <fragment
-        android:id="@+id/gMap"
-        android:name="com.google.android.gms.maps.SupportMapFragment"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toTopOf="@+id/city_bikes"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:context=".GMapsActivity" />
+    <Button
+        android:id="@+id/city_bikes"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="36dp"
+        android:text="@string/city_bikes"
+        app:layout_constraintBaseline_toBaselineOf="@+id/air_stations"
+        app:layout_constraintStart_toStartOf="parent" />
 
     <Button
         android:id="@+id/air_stations"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginEnd="8dp"
-        android:layout_marginStart="8dp"
-        android:text="@string/air_stations"
+        android:layout_marginEnd="18dp"
+        android:layout_marginTop="93dp"
         android:onClick="airStationsAPIActivity"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:text="@string/air_stations"
         app:layout_constraintEnd_toStartOf="@+id/parking"
-        app:layout_constraintHorizontal_bias="0.795"
-        app:layout_constraintStart_toEndOf="@+id/city_bikes" />
-
-    <Button
-        android:id="@+id/city_bikes"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/city_bikes"
-        app:layout_constraintBottom_toBottomOf="parent" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <Button
         android:id="@+id/parking"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="280dp"
         android:onClick="parkingAPIActivity"
         android:text="@string/parking"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
+        app:layout_constraintBaseline_toBaselineOf="@+id/air_stations"
+        app:layout_constraintStart_toStartOf="parent" />
 
+
+    <Button
+        android:id="@+id/map"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="118dp"
+        android:onClick="viewMap"
+        android:text="@string/map"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 </android.support.constraint.ConstraintLayout>
